toggle-wrapper: fix issue with single child
If the toggle applies only on 1 child field, the behavior of the toggle failed: the child field was not hided.
The problem was that we use `this.model` to check if children elements are empty. The behavior of this model
property change if the element are one or multiple children :
- if one child : `this.model` contains all documents values (and so `checkIsEmpty` return most of the time False
- if two or more childrend : `this.model` contains only children elements (ans so `checkIsEmpty` is relevant)

Yo solve ythis problem, we now use the value returne by `valueChanges` observer to determine if child/chilren
are empty or not.

Co-Authored-by: Renaud Michotte <renaud.michotte@gmail.com>