editor: fix preprocessRecord calls
This PR fixes the redundant calls to 'preprocessRecord'. Before this PR
the preprocessRecord function was called each time the model was getted.
This will cause a call for each field present into the editor. This PR
fixes this problem calling preprocessRecord function only when the model
is setted or when it changes from the parent input.

Co-Authored-by: Johnny Mariéthoz <Johnny.Mariethoz@rero.ch>
Co-Authored-by: Renaud Michotte <renaud.michotte@gmail.com>