editor: fix problem with 'jump to' component
Some entries from 'Jump To' components wasn't working as expected : the
focus isn't set into the first visible field of the chosen entry. This
behavior was cause by hidden fields : if the first input field for the
chosen entry is an hidden field, then the focus was set on it... This
commit solved this problem by setting the focus on the first visible
input field of the chosen 'Jump To' entry.

* Closes rero/rero-ils#1035

Ca-Authored-By: Renaud Michotte <renaud.michotte@gmail.com>