menu: add Id option for menu and submenu
Cypress tests reveals a difficulty to target some menu. If you're in
different languages you cannot use an item description (as
"User"/"Utilisateur").
Cypress gives the possibility to use "id=" in html content to target a
specific item in the menu.
This commit improves menu component to also display "id=" in html
content if the menu entry gives an id parameter.

* Adds Id option for menu

Co-Authored-by: Bertrand Zuchuat <bertrand.zuchuat@rero.ch>
Co-Authored-by: Alicia Zangger <alicia.zangger@rero.ch>
Co-Authored-by: Olivier DOSSMANN <git@dossmann.net>