editor: fix hidden field with value from the model
Some fields containing values comming from the input model should be visible
after the editor initialization. This is not the case with a recent version of
ngx formly. This is due to the use of bad hack `setTimeout` at several
places in the code.  As ngx formly made the think cleaner these hacks have been
removed.

* Cleans hidden field list at the component destruction instead of
initialization.
* Reloads the JSONSchema once the model has changed.
* Rewrites the add field component by using observers only.

Co-Authored-by: Johnny Mariéthoz <Johnny.Mariethoz@rero.ch>
Co-Authored-by: Sébastien Délèze <sebastien.deleze@rero.ch>