editor: fix ToggleSwitch & DatePicker
At this time, the DatePicker component reset function doesn't work as
expected. When the reset() function is called, the value is set to
1970-01-01 (zero epoch timestamp) instead to set an empty string into
the field. This commit fixes this wrong behavior.

This commit also fixes a problem from the ToggleSwitch component. The
toggle switch was not correctly initialized if it is related to a simple
component (<> Array <> Object)

Co-Authored-by: Renaud Michotte <renaud.michotte@gmail.com>
Co-Authored-by: Johnny Mariéthoz <Johnny.Mariethoz@rero.ch>