template: allow using templates with new holdings/items
* When we create a new holding/item from 'scratch', it will be linked
with a document. This is done by adding a 'document' argument into the
query url. Using a template removes this argument from query string and
cause an HTTP 400 error when trying to save because the editor lost the
link with the document. This commit solves this problem keeping all
existing arguments from query string when loading a template.

Co-authored-by: Renaud Michotte <renaud.michotte@gmail.com>