views: add button to use a resource
This commit add the possibility to use a resource. If a resource can be
used, it is associated to an URL to use this resource. This commit also
add a button ('use') for each search result and in the detail view. This
button can be controlled by the 'canUse' parameter from the resource
configuration.

Note :
  * 'editing a resource' is the fact to open the editor for this resource
and update the data.
  * 'using a resource' if the fact to use the resource as a template
skeleton to create an other resource (not necessarily the same resource
type).

Co-authored-by: Renaud Michotte <renaud.michotte@gmail.com>