editor: fix cancel button behavior.
When we use the "cancel" button of the editor, the basic behavior is to
navigate to the previous URL. But, when we use the load template function,
this behavior results in reloading the form (without templates URL
parameters) and one needs to click again on "cancel" button.

This commit fixes this problem by skipping the "loading template URL"
from the browser history.

Closes rero/rero-ils#1453

Authored-by: Renaud Michotte <renaud.michotte@gmail.com>