bug: correct dropdown menu issue and modal tester config.
In a menu, when a dropdown entry was clicked, the page was reloaded. It
caused by an empty `href` tag. Using an `href` tag  without any value
and a 'click' event property binding (as mentioned in the official doc :
https://valor-software.com/ngx-bootstrap/#/dropdowns#anchor-trigger),
all seems work much better.

In the app tester, the button configuration for the modal actions was
missing. Additionally, add default values for modal button if
configuration doesn't provide any values.

Co-Authored-by: Renaud Michotte <renaud.michotte@gmail.com>