search: fix number facets behavior
Behavior to know if a facet is selected or not are based on query string
argument; the value of a query string argument is always a "string". But
For an aggregation/facet with only numeric values, all checks failed
because we compared "string" versus "number". To fix this, we add the
".toString()" (available for all native datatype) method into the
 required functions.

Co-authored-by: Renaud Michotte <renaud.michotte@gmail.com>